@@ -4,39 +4,57 @@
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build"
+    "build": "vue-cli-service build",
+    "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "@editorjs/editorjs": "^2.18.0",
-    "@vue/composition-api": "^0.5.0",
-    "firebase": "^9.9.3",
-    "vue-editor-js": "^2.0.2",
-    "vue-select": "^3.20.0",
-    "vuedraggable": "^2.24.3",
+    "@editorjs/checklist": "^1.4.0",
+    "@editorjs/code": "^2.8.0",
+    "@editorjs/delimiter": "^1.3.0",
+    "@editorjs/embed": "^2.5.3",
+    "@editorjs/header": "^2.7.0",
+    "@editorjs/inline-code": "^1.4.0",
+    "@editorjs/list": "^1.8.0",
+    "@editorjs/marker": "^1.3.0",
+    "@editorjs/paragraph": "^2.9.0",
+    "@editorjs/table": "^2.2.1",
+    "@vueuse/core": "^9.13.0",
     "core-js": "^3.8.3",
-    "vue": "2.6.11",
-    "@editorjs/checklist": "^1.1.0",
-    "@editorjs/code": "^2.5.0",
-    "@editorjs/delimiter": "^1.1.0",
-    "@editorjs/embed": "^2.3.1",
-    "@editorjs/header": "^2.4.1",
-    "@editorjs/image": "^2.4.1",
-    "@editorjs/inline-code": "^1.3.1",
-    "@editorjs/link": "^2.2.0",
-    "@editorjs/list": "^1.4.0",
-    "@editorjs/marker": "^1.2.2",
-    "@editorjs/paragraph": "^2.6.1",
-    "@editorjs/personality": "^2.0.1",
-    "@editorjs/quote": "^2.3.0",
-    "@editorjs/raw": "^2.1.2",
-    "@editorjs/simple-image": "^1.3.3",
-    "@editorjs/table": "^1.2.2",
-    "@editorjs/warning": "^1.1.1"
+    "pinia": "^2.0.32",
+    "uuid": "^9.0.0",
+    "vue": "^3.2.13",
+    "vue-editor-js": "^2.0.2",
+    "vue-select": "^4.0.0-beta.6"
   },
   "devDependencies": {
     "@babel/core": "^7.12.16",
+    "@babel/eslint-parser": "^7.12.16",
+    "@editorjs/personality": "^2.0.2",
+    "@types/editorjs__header": "^2.6.0",
     "@vue/cli-plugin-babel": "~5.0.0",
+    "@vue/cli-plugin-eslint": "~5.0.0",
     "@vue/cli-service": "~5.0.0",
-    "vue-template-compiler": "2.6.11"
-  }
+    "eslint": "^7.32.0",
+    "eslint-plugin-vue": "^8.0.3"
+  },
+  "eslintConfig": {
+    "root": true,
+    "env": {
+      "node": true
+    },
+    "extends": [
+      "plugin:vue/vue3-essential",
+      "eslint:recommended"
+    ],
+    "parserOptions": {
+      "parser": "@babel/eslint-parser"
+    },
+    "rules": {}
+  },
+  "browserslist": [
+    "> 1%",
+    "last 2 versions",
+    "not dead",
+    "not ie 11"
+  ]
 }
