@@ -1,894 +1,30 @@
 <template>
-  <div class="page">
-    <div v-if="pageReady">
-      <nav class="site-nav">
-        <ul>
-          <li v-if="0"><button type="button" class="muted" @click="invokeSave"><img src="./assets/images/cloud.svg" :class="saving ? 'muted' : ''" alt="Autosave" aria-label="Autosaving" /></button></li>
-          <li v-if="0"><button type="button" :class="authenticated ? '' : 'muted'" @click="modal = 'account'"><img src="./assets/images/account.svg" alt="Account" aria-label="Your Account" /></button></li>
-          <li><button class="muted" type="button" @click="modal = 'about'" target="_blank"><img src="./assets/images/dots.svg" alt="More" aria-label="About this site"></button></li>
-        </ul>
-      </nav>
-      <main class="container">
-        <input type="text" class="task-title" v-model="title" placeholder="Name your task" />
-        <div style="display: flex; margin: 20px 0; align-center: center; gap: 10px">
-          <img v-if="!lightningMode" src="./assets/images/time.svg" alt="" />
-          <img v-else src="./assets/images/lightning-mode.svg" alt="Lightning mode" />
-          <v-select :disabled="working" v-model="seconds" style="width: 50%; border: 0" :options="timeOptions"></v-select>
-        </div>
-        <editor @ready="loadEditor()" ref="editor" :config="config" />
-      </main>
-      <footer class="controls">
-        <ul class="folders">
-          <li>
-            <button v-if="authenticated && !premium" style="display: inline-flex; align-items: center; gap: 10px;"  type="button" @click="queueOpen = !queueOpen"><img src="./assets/images/queue.svg" alt="Folder" /></button>
-            <div class="queue-list" v-if="queueOpen">
-              <div>
-                <div v-if="shuffle">
-                  <div style="border-bottom: #ddd 1px solid; margin-bottom: 5px; display: flex; align-items: center; justify-content: space-between" v-for="shuffleTask in shuffledTasks" :key="'shuffle-'+shuffleTask.id">
-                    <span @click="changeTask(shuffleTask)" :style="shuffleTask.completed ? {fontSize: '15px', padding: '5px 3px', textDecoration: 'line-through', opacity: .5} : {fontSize: '15px', padding: '5px 3px'}">{{shuffleTask.title}}</span>
-                    <button v-if="tasks.length > 1 && id !== shuffleTask.id" class="muted" style="font-size: 1.25rem" type="button" @click="modal = 'deleteTask'; selectedTask = shuffleTask" aria-label="Delete task"><img src="./assets/images/trash.svg" alt="Delete" /></button>
-                  </div>
-                </div>
-                <draggable v-else :list="tasks" group="tasks-todo">
-                  <div style="border-bottom: #ddd 1px solid; margin-bottom: 5px; display: flex; align-items: center; justify-content: space-between" v-for="task in tasks" :key="task.id">
-                    <span @click="changeTask(task)" :style="task.completed ? {fontSize: '15px', padding: '5px 3px', textDecoration: 'line-through', opacity: .5} : {fontSize: '15px', padding: '5px 3px'}">{{task.title}}</span>
-                    <button v-if="tasks.length > 1 && taskId !== task.id" class="muted" style="font-size: 1.25rem" type="button" @click="modal = 'deleteTask'; selectedTask = task" aria-label="Delete task"><img src="./assets/images/trash.svg" alt="Delete" /></button>
-                  </div>
-                </draggable>
-              </div>
-              <div>
-                <form @submit.prevent="addNewTask(newTaskTitle)" class="add-new-task">
-                  <input style="width: 100%" required v-model="newTaskTitle" type="text" placeholder="New task" />
-                  <button :class="newTaskTitle.trim() === '' ? 'invisible muted' : 'invisible'" :disabled="newTaskTitle.trim() === ''" type="submit" aria-label="Create new task"><img src="./assets/images/plus.svg" alt="Add" /></button>
-                </form>
-              </div>
-            </div>
-            <button v-if="authenticated && premium" style="display: inline-flex; align-items: center; gap: 4px;" @click="foldersOpen = !foldersOpen" class="folders-select" type="button"><img src="./assets/images/folder.svg" alt="Folder" /> {{currentFolder.name}}</button>
-            <button v-if="!authenticated" style="display: inline-flex; align-items: center; gap: 10px;"  class="folders-select" type="button"><img src="./assets/images/queue.svg" alt="Folder" /> <span class="queue-login">Queue in future</span></button>
-            <div v-if="foldersOpen" class="folders-list">
-              <draggable group="folders" :list="folders">
-                <div style="border-bottom: #ddd solid 1px; margin-bottom: 5px; padding-bottom: 5px; display: flex; align-items: flex-start; justify-content: space-between" v-for="folder in folders" :key="folder.id">
-                  <button style="text-align: left" class="change-folder" type="button" @click="changeFolder(folder)" :aria-label="'Change to '+folder.name+' folder'">{{folder.name}}</button>
-                  <div style="display: flex; gap: 5px;">
-                    <button class="muted" type="button" :aria-label="'Rename '+folder.name" @click="modal = 'renameFolder'; selectedFolder = folder">
-                      <img src="./assets/images/pencil.svg" />
-                    </button>
-                    <button v-if="justCopied === folder.id" class="muted" @blur="justCopied = false" type="button" :aria-label="'Copy '+folder.name" @click="changeFolder(folder)">
-                      <img src="./assets/images/arrow.svg" />
-                    </button>
-                    <button v-else class="muted" @blur="justCopied = false" type="button" :aria-label="'Copy '+folder.name" @click="copyFolder(folder)">
-                      <img src="./assets/images/copy.svg" alt="Copy" />
-                    </button>
-                    <button v-if="folders.length > 1 && currentFolder.id !== folder.id" class="muted" type="button" :aria-label="'Delete '+folder.name" @click="modal = 'deleteFolder'; selectedFolder = folder"><img src="./assets/images/trash.svg" alt="Trash" /></button>
-                  </div>
-                </div>
-              </draggable>
-              <form @submit.prevent="addNewFolder(newFolderName)" class="add-new-folder">
-                <input style="width: 100%" v-model="newFolderName" type="text" placeholder="New folder name" />
-                <button class="invisible" ref="addNewFolderButton" :class="newFolderName.trim() === '' ? 'muted' : ''" :disabled="newFolderName.trim() === ''" type="submit"  aria-label="Create new folder"><img src="./assets/images/plus.svg" alt="Add" /></button>
-              </form>
-            </div>
-          </li>
-        </ul>
-        <div class="action">
-          <div class="tooltip">
-            <button type="button" :class="lightningMode ? '' : 'muted'" @click="lightningMode = !lightningMode" v-if="authenticated">
-              <img src="./assets/images/lightning.svg" alt="Lightning Mode" aria-label="Halves the remaining time" />
-            </button>
-            <span class="tooltiptext">Lightning</span>
-          </div>
-          <div class="tooltip">
-            <button type="button" v-if="authenticated" @click="shuffle = !shuffle" :class="shuffle ? '' : 'muted'"><img src="./assets/images/shuffle.svg" alt="Shuffle" aria-label="Shuffle tasks" /></button>
-            <span class="tooltiptext">Shuffle</span>
-          </div>
-          <div class="browse">
-            <div class="muted tooltip">
-              <!-- <button type="button"><img src="./assets/images/startover.svg" alt="Start over" aria-label="Start over" /></button> -->
-              <span class="tooltiptext">Start over</span>
-            </div>
-            <div>
-              <button v-if="working" type="button" @click="pause()"><img src="./assets/images/pause.svg" alt="Pause" aria-label="Pause task" /></button>
-              <button v-else type="button"  @click="play()"><img src="./assets/images/play.svg" alt="Start" aria-label="Start task" /></button>
-            </div>
-            <div class="muted tooltip">
-              <!-- <button @click="authenticated ? goToNextTask() : modal = 'account'" type="button"><img src="./assets/images/skip.svg" alt="Skip" aria-label="Go to next task" /></button> -->
-              <span class="tooltiptext">Next task</span>
-            </div>
-          </div>
-          <div class="tooltip">
-            <div v-if="authenticated">
-              <button v-if="autoplay" @click="autoplay = !autoplay" type="button"><img src="./assets/images/autoplay.svg" alt="Autoplay" aria-label="Start next task automatically" /></button>
-              <button v-else @click="autoplay = !autoplay" type="button" class="muted"><img src="./assets/images/autoplay-off.svg" alt="Autoplay" aria-label="Don’t next task automatically" /></button>
-            </div>
-            <span class="tooltiptext">Autoplay</span>
-          </div>
-            <ul class="folders">
-              <li v-if="authenticated && premium">
-                <button @click="queueOpen = !queueOpen" type="button">
-                  <img src="./assets/images/queue.svg" alt="List" />
-                </button>
-            </li>
-          </ul>
-        </div>
-        <div class="mark-complete">
-          <button @click="completed = false" v-if="completed" type="button"><img src="./assets/images/checkmark-done.svg" alt="Yay!" aria-label="All done!" /></button>
-          <button @click="skip(true)" v-else type="button"><img src="./assets/images/mark-complete.svg" alt="Mark Complete" aria-label="All done!" /></button>
-        </div>
-      </footer>
-      <div @click="goToNextColor()" style="width: 30px" class="color" :style="{ backgroundColor: activeColor }"></div>
-      <div @click="goToNextColor()" class="progress-bar">
-        <div @click="goToNextColor()" class="progress-bar-color" :style="{width: progressBarWidth, backgroundColor: activeColor, transitionDuration: transitionDuration}"></div>
-      </div>
-      <div @click="foldersOpen = false; queueOpen = false" class="screen" v-if="foldersOpen || queueOpen"></div>
-      <div @click="modal = false; waitingForSignin = false" class="overlay" v-if="modal"></div>
-      <div class="modal" v-if="modal === 'account'">
-        <button @click="modal = false; waitingForSignin = false" type="button" class="close-button">&times;</button>
-        <div v-if="authenticated">
-          <h2>Your account</h2>
-          <p>Email: {{email}}</p>
-          <p style="padding-bottom: 10px; display: flex; gap: 10px"><button @click="modal = false" type="submit">OK</button><button @click="logout" type="button">Log out</button></p>
-        </div>
-        <div v-else>
-          <h2>Your account</h2>
-          <p>Log into your existing account or create a new one if none exists.</p>
-          <p style="font-size: 16px">Wondering why you should register? Add a queue and access your notes/preferences across devices for free.</p>
-          <form @submit.prevent="sendEmailSignIn">
-            <p><label>Email
-              <input :disabled="waitingForSignin" v-model="email" type="email" required />
-            </label></p>
-            <p style="padding-bottom: 10px"><button :disabled="!email" :class="waitingForSignin || !email ? 'muted' : ''" type="submit">{{waitingForSignin ? 'Check your email' : 'Register / login'}}</button><br><span style="display: block; margin-top: 10px; font-size: 16px">An instant sign-in link will be emailed to you.</span></p>
-          </form>
-        </div>
-      </div>
-      <div class="modal" v-if="modal === 'premium'">
-        <button @click="modal = false" type="button" class="close-button">&times;</button>
-        <div v-if="premium">
-          <h2>You’re premium!</h2>
-          <ul class="features">
-            <li>Cloneable folders</li>
-            <li>Task archives</li>
-            <li>Zapier integration</li>
-          </ul>
-          <p>Want to cancel or manage your subscription?</p>
-          <form @submit.prevent="goPremium(false); modal = false">
-            <p style="padding-bottom: 10px; font-size: 15px;"><button type="submit">Go to manage</button></p>
-          </form>
-        </div>
-        <div v-else>
-          <h2>Go premium</h2>
-          <p>Knock out your todo lists for just $5 USD monthly.</p>
-          <ul class="features">
-            <li>Cloneable tasks</li>
-            <li>Task archives</li>
-            <li>Zapier integration</li>
-          </ul>
-          <form @submit.prevent="goPremium(); modal = false">
-            <p style="padding-bottom: 10px; line-height: 2; font-size: 15px;"><button type="submit">Subscribe</button><br>You can cancel anytime!</p>
-          </form>
-        </div>
-      </div>
-      <div class="modal" v-if="modal === 'restart'">
-        <button @click="modal = false" type="button" class="close-button">&times;</button>
-        <h2>Start task over?</h2>
-        <p>This will bring the task to the top of the queue.</p>
-        <form>
-          <p style="padding-bottom: 10px; line-height: 2; font-size: 15px; display: inline-flex; gap: 10px"><button type="submit">Yes</button> <button type="button" @click="modal = false">No</button></p>
-        </form>
-      </div>
-      <div class="modal" v-if="modal === 'deleteFolder'">
-        <button @click="modal = false" type="button" class="close-button">&times;</button>
-        <h2>Delete {{selectedFolder.name}}?</h2>
-        <p>This is irreversible!</p>
-        <form>
-          <p style="padding-bottom: 10px; line-height: 2; font-size: 15px; display: inline-flex; gap: 10px"><button type="button" class="submit" @click="modal = false">Never mind</button><button type="button" class="invisible" @click="deleteFolder(selectedFolder.id); modal = false">Delete</button></p>
-        </form>
-      </div>
-      <div class="modal" v-if="modal === 'deleteTask'">
-        <button @click="modal = false" type="button" class="close-button">&times;</button>
-        <h2>Delete {{selectedTask.title}}?</h2>
-        <p>This is irreversible! Notes will be gone.</p>
-        <p style="padding-bottom: 10px; line-height: 2; font-size: 15px; display: inline-flex; gap: 10px"><button type="button" class="submit" @click="modal = false">Never mind</button><button type="button" class="invisible" @click="deleteTask(selectedTask); modal = false">Delete</button></p>
-      </div>
-      <div class="modal" v-if="modal === 'renameFolder'">
-        <button @click="modal = false" type="button" class="close-button">&times;</button>
-        <h2>Rename <code>{{selectedFolder.name}}</code></h2>
-        <form @submit.prevent="renameFolder(selectedFolder); modal = false">
-          <p><label>New name<br>
-            <input v-model="renameFolderName" style="font: inherit" type="text" class="show" required />
-          </label></p>
-          <p style="padding-bottom: 10px; line-height: 2; font-size: 15px; display: inline-flex; gap: 10px"><button type="submit">Ready</button><button type="button" class="" @click="modal = false">Never mind</button></p>
-        </form>
-      </div>
-      <div class="modal" v-if="modal === 'moreTime'">
-        <button @click="modal = false" type="button" class="close-button">&times;</button>
-        <h2>Do you want to add another {{seconds.label}}?</h2>
-        <p style="padding-bottom: 10px; line-height: 2; font-size: 15px; display: inline-flex; gap: 10px"><button class="submit" @click="startOver(); modal = false" type="button">Yes</button><button type="button" class="" @click="modal = false">No</button></p>
-      </div>
-      <div class="modal" v-if="modal === 'about'" style="font-size: 18px">
-        <button @click="modal = false" type="button" class="close-button">&times;</button>
-        <h2>Hey, I’m Amí ✌🏼</h2>
-        <p>I’m a <a href="https://naeily.com" target="_blank">freelance web designer and developer</a> based in New Mexico. I built Focusmix with Vue.js, Firebase, and Editor.js.</p>
-        <p>Focusmix is in <strong>ALPHA</strong>: Do not use for anything important for now. <a href="https://subscribepage.io/BNo1VG" target="_blank">Sign up for email updates</a>.</p>
-        <p>Need support? Email <a href="mailto:ami@naeily.com">ami@naeily.com</a>.</p>
-        <form>
-          <p style="padding-bottom: 10px; line-height: 2; font-size: 15px; display: inline-flex; gap: 10px"><button @click="modal = false" type="submit">OK</button></p>
-        </form>
-      </div>
-    </div>
-    <div v-else class="loading">Loading your experience...</div>
+  <user-nav />
+  <main-content />
+  <controls-nav />
+  <div class="modals">
+    <about-site-modal />
+    <queue-modal />
+    <task-over-modal />
   </div>
 </template>
 
 <script>
-import Header from '@editorjs/header';
-import List from '@editorjs/list';
-import CodeTool from '@editorjs/code'
-import Paragraph from '@editorjs/paragraph'
-import Embed from '@editorjs/embed'
-import Table from '@editorjs/table'
-import Checklist from '@editorjs/checklist'
-import Marker from '@editorjs/marker'
-import Delimiter from '@editorjs/delimiter'
-import InlineCode from '@editorjs/inline-code';
-import draggable from "vuedraggable";
-import {firebase, db, auth} from './db'
-import { v4 as uuidv4 } from 'uuid';
-import vSelect from 'vue-select'
-import 'vue-select/dist/vue-select.css';
-import Times from './assets/json/times.json'
+import UserNav from '@/components/UserNav'
+import MainContent from '@/components/MainContent'
+import ControlsNav from '@/components/ControlsNav'
+import AboutSiteModal from '@/components/ModalPopups/AboutSiteModal.vue'
+import QueueModal from '@/components/ModalPopups/QueueModal.vue'
+import TaskOverModal from '@/components/ModalPopups/TaskOverModal.vue'
 
 export default {
-  name: 'App',
-  components: { draggable, vSelect },
-  data() {
-    return {
-      selectedTask: null,
-      timeLeft: 3600,
-      pageReady: false,
-      activeColor: '#ffffff',
-      progressBarWidth: '1px',
-      colors: [
-        '#ffe8fa',
-        '#fdd2d1',
-        '#fff2ce',
-        '#ebf4e1',
-        '#eaf9ff',
-        '#f4edff',
-        '#f2f2f2'
-      ],
-      timer: null,
-      timeOptions: Times,
-      tasks: [],
-      shuffledTasks: [],
-      folders: [],
-      justCopied: false,
-      config: {
-        tools: {
-          header: {
-            class: Header,
-            config: {
-              placeholder: 'Enter a header',
-              levels: [2, 3, 4],
-              defaultLevel: 1,
-            }
-          },
-          list: {
-            class: List,
-            inlineToolbar: true,
-          },
-          code: {
-            class: CodeTool
-          },
-          paragraph: {
-            class: Paragraph,
-          },
-          embed: {
-            class: Embed,
-            config: {
-              services: {
-                youtube: true,
-                imgur: true,
-                twitter: true
-              }
-            }
-          },
-          table: {
-            class: Table,
-            inlineToolbar: true,
-            config: {
-              rows: 2,
-              cols: 3,
-            },
-          },
-          checklist: {
-            class: Checklist,
-          },
-          Marker: {
-            class: Marker,
-            shortcut: 'CMD+SHIFT+M',
-          },
-          inlineCode: {
-            class: InlineCode,
-            shortcut: 'CMD+SHIFT+C',
-          },
-          delimiter: Delimiter,
-        }
-      },
-      modal: false,
-      clear: false,
-      newFolderName: '',
-      newTaskTitle: '',
-      authenticated: false,
-      premium: false,
-      lightningMode: false,
-      shuffle: false,
-      autoplay: false,
-      working: false,
-      foldersOpen: false,
-      queue: [],
-      queueOpen: false,
-      modal: false,
-      selectedFolder: null,
-      selectedTask: null,
-      email: '',
-      waitingForSignin: false,
-      uid: '',
-      saving: false,
-      title: '',
-      seconds: {
-        label: '1 hour',
-        value: 3600,
-      },
-      notes: {},
-      elapsed: 0,
-      completed: false,
-      currentFolder: {id: ''},
-      renameFolderName: '',
-      transitionDuration: '0s, 100ms, 100ms',
-      beep: require('./assets/media/beep.mp3'),
-      success: require('./assets/media/success.mp3'),
-      welcomeBlocks: [
-        {
-          data: {
-            text: 'Welcome. Focusmix is a tool for getting stuff done.'
-          },
-          id: "moTtRFW4VL",
-          type: "paragraph"
-        },
-        {
-          data: {text: '▶️ Hit the play button.'},
-          id: "xGAIubcQNS",
-          type: "paragraph"
-        },
-        {
-          data: {text: 'As you’ll see, the page is a progress bar. (Click the color bar the left side to change to your favorite.)'},
-          id: "HcSf6A18HS",
-          type: "paragraph"
-        },
-        {
-          data: {text: 'This area you’re reading is for notes while you’re working.'},
-          id: "l8gu-1k-6S",
-          type: "paragraph"
-        },
-        {
-          data: {text: 'Free features behind a login include a queue and archive, lightning mode ⚡️ (halves the remaining time), shuffle, autoplay, and cross-device access.'},
-          id: "seYBLpHl2W",
-          type: "paragraph"
-        },
-        {
-          data: {text: 'Enjoy!'},
-          id: "sQWPzD4NOz",
-          type: "paragraph"
-        }
-      ]
-    }
-  },
-  mounted() {
-    var getQueryVariable = (variable) => {
-      var query = window.location.search.substring(1);
-      var vars = query.split("&");
-      for (var i=0;i<vars.length;i++) {
-        var pair = vars[i].split("=");
-        if(pair[0] == variable){return pair[1];}
-      }
-      return(false);
-    }
-
-    if (getQueryVariable('mode') === 'signIn') {
-      auth.fetchSignInMethodsForEmail(localStorage.getItem('emailForSignIn')).then((methods) => {
-        if (auth.isSignInWithEmailLink(window.location.href)) {
-          const email = localStorage.getItem('emailForSignIn')
-          if (!email) {
-            email = window.prompt('Please provide your email for confirmation.')
-          } if (email) {
-            auth.signInWithEmailLink(email, window.location.href)
-            .then(() => {
-              this.authenticated = true;
-              this.modal = 'account';
-              this.email = email
-              this.uid = auth.currentUser.uid
-              localStorage.removeItem('emailForSignIn');
-              window.history.pushState({}, document.title, window.location.pathname);
-              setTimeout(() => {
-                if (methods.length === 0) {
-                  this.convertAccount(this.uid)
-                } else {
-                  this.pullData()
-                }
-              }, 888)
-            });
-          } else {
-            alert('Try again.')
-          }
-        }
-      })
-      setTimeout(() => {
-        this.invokeSave()
-        setInterval(() => {
-          this.invokeSave()
-        }, 5555)
-      }, 9999)
-    } else {
-      auth.onAuthStateChanged((user) => {
-        if (user) {
-          this.uid = user.uid
-          this.email = user.email
-          this.authenticated = true
-          this.pullData();
-          setTimeout(() => {
-            this.pageReady = true
-          }, 888)
-        } else {
-          this.uid = false
-          const randomColor = this.colors[Math.floor(Math.random() * this.colors.length)];
-          this.activeColor = randomColor
-          if (Boolean(parseInt(this.getLocalStorage('setupComplete')))) {
-            this.pullLocalData()
-          } else {
-            window.localStorage.clear()
-            this.createWelcome()
-          }
-
-          setTimeout(() => {
-            this.pageReady = true
-          }, 888)
-        }
-        setInterval(() => {
-          this.invokeSave()
-        }, 5555)
-      });
-    }
-  },
-  watch: {
-    seconds(newValue) {
-      if (this.autoplay) {
-        this.transitionDuration = newValue.value-this.elapsed+'s, 100ms, 100ms'
-      } else {
-        this.transitionDuration = '100ms, 100ms, 100ms'
-      }
-    },
-    title(newValue, previousValue){
-      if (!this.authenticated) return;
-      if (newValue !== previousValue) {
-        this.setLocalStorage('title', newValue)
-        var taskIndex = this.tasks.indexOf(this.tasks.find(element => element.id === this.taskId))
-        var original = this.tasks
-        var copy = [].concat(original);
-        if (copy[taskIndex] && copy[taskIndex].title) {
-          copy[taskIndex].title = newValue
-        }
-        this.tasks = copy
-      }
-    },
-    seconds(newValue) {
-      this.transitionDuration = '100ms, 100ms, 100ms'
-    },
-    lightningMode(newValue) {
-      if (this.working) {
-        if (newValue) {
-          this.transitionDuration = (this.seconds.value-this.elapsed)/2+'s, 100ms, 100ms'
-          this.progressBarWidth = 'calc(100% - 30px)'
-        } else {
-          this.transitionDuration = this.seconds.value-this.elapsed+'s, 100ms, 100ms'
-          this.progressBarWidth = 'calc(99.99% - 30px)'
-        }
-      }
-    },
-    working(newValue) {
-      if (!newValue) this.transitionDuration = '0ms, 100ms, 100ms'
-    },
-    shuffle(enabled) {
-      if (enabled) {
-        this.shuffledTasks = this.shuffleTasks()
-      }
-    },
-  },
-  created() {
-    window.addEventListener('beforeunload', this.beforeWindowUnload)
-  },
-  beforeDestroy() {
-    window.removeEventListener('beforeunload', this.beforeWindowUnload)
-  },
-  methods: {
-    goToNextTask() {
-      this.pause()
-      var tasks = this.tasks
-      if (this.shuffle) {
-        tasks = this.shuffledTasks
-      }
-      var thisTask = tasks.indexOf(tasks.find(element => element.id === this.taskId))
-      var nextTask = this.getNextArrItem(thisTask, tasks)
-
-      if (nextTask) {
-        this.changeTask(nextTask)
-        if (this.autoplay) {
-          this.play()
-        }
-      } else {
-        this.addNewTask('New task', true)
-      }
-    },
-    deleteTask(task) {
-      this.saving = true;
-      db.collection('Users').doc(this.uid).collection('Tasks').doc(task.id).delete().then(() => {
-        var selectedTaskIndex = this.tasks.indexOf(this.tasks.find(element => element.id === task.id))
-        var original = this.tasks
-        var newTasks = [].concat(original);
-        newTasks.splice(selectedTaskIndex, 1)
-        this.tasks = newTasks
-        db.collection('Users').doc(this.uid).collection('Folders').doc(this.currentFolder.id).delete().then(() => {
-          db.collection('Users').doc(this.uid).collection('Folders').doc(this.currentFolder.id).set({
-            tasks: this.tasks
-          }, { merge: true }).then(() => {
-            this.saving = false;
-          })
-        });
-      }).catch((err) => {
-        console.log(err)
-      });
-    },
-    changeTask(task) {
-      this.pause()
-      db.collection('Users').doc(this.uid).collection('Tasks').doc(task.id).get().then((doc) => {
-        this.taskId = task.id
-        this.title = doc.data().title
-        this.seconds = doc.data().seconds
-        this.elapsed = doc.data().elapsed
-        this.transitionDuration = '0.001s, 100ms, 100ms'
-        this.progressBarWidth = 'calc('+this.elapsed/this.seconds.value*100+'vw - 30px)'
-        this.completed = doc.data().completed
-        this.notes = doc.data().notes
-        this.$refs.editor._data.state.editor.render(doc.data().notes)
-        db.collection('Users').doc(this.uid).get((doc) => {
-          db.collection('Users').doc(this.uid).set({
-            workingTask: task.id
-            }, { merge: true })
-        })
-      })
-    },
-    pullLocalData() {
-      this.title = this.getLocalStorage('title')
-      this.activeColor = this.getLocalStorage('activeColor')
-      this.seconds = JSON.parse(this.getLocalStorage('seconds'))
-      this.notes = JSON.parse(this.getLocalStorage('notes'))
-      this.completed = Boolean(parseInt(this.getLocalStorage('completed')))
-      this.activeColor = this.getLocalStorage('color')
-      this.elapsed = parseInt(this.getLocalStorage('elapsed'))
-      this.taskId = this.getLocalStorage('taskId')
-      this.config.data = this.notes
-    },
-    skip(completed = false) {
-      if (completed) {
-        this.completed = true
-        // if (this.authenticated) {
-        //   if (this.tasks.length > 1) {
-        //   }
-        // } else {
-        //   this.modal = 'account'
-        // }
-        var audio = new Audio(this.success);
-        audio.play();
-        this.pause()
-      } else {
-        var audio = new Audio(this.beep);
-        this.modal = 'moreTime'
-        this.pause()
-        audio.play();
-      }
-    },
-    convertAccount(uid) {
-      var onlyFolder = { id: uuidv4(), name: 'Folder' }
-      this.title = this.getLocalStorage('title')
-      this.notes = JSON.parse(this.getLocalStorage('notes'))
-      this.seconds = JSON.parse(this.getLocalStorage('seconds'))
-      this.elapsed = parseInt(this.getLocalStorage('elapsed'))
-      this.completed = Boolean(parseInt(this.getLocalStorage('completed')))
-      this.taskId = uuidv4()
-      this.activeColor = this.getLocalStorage('color')
-      this.currentFolder = onlyFolder
-      window.localStorage.clear()
-
-      db.collection('Users').doc(uid).set({
-        options: {
-          color: this.activeColor,
-        },
-        workingFolder: this.currentFolder,
-        workingTask: this.taskId,
-        folders: [onlyFolder]
-      })
-
-      db.collection('Users').doc(uid).collection('Tasks').doc(this.taskId).set({
-        notes: this.notes,
-        title: this.title,
-        seconds: this.seconds,
-        elapsed: this.elapsed,
-        completed: this.completed
-      })
-
-      db.collection('Users').doc(uid).collection('Folders').doc(this.currentFolder.id).set({
-        tasks: [
-          {
-            id: this.taskId,
-            title: this.title,
-            completed: this.completed
-          }
-        ],
-      })
-
-      setTimeout(() => {
-        window.location.reload()
-      }, 1111)
-    },
-    toTimeString(seconds) {
-      return new Date(seconds * 1000).toISOString().substr(11, 8);;
-    },
-    play() {
-      this.completed = false;
-      this.progressBarWidth = 'calc(100vw - 30px)'
-      if (this.lightningMode) {
-        this.transitionDuration = (this.seconds.value-this.elapsed)/2+'s, 100ms, 100ms'
-      } else {
-        this.transitionDuration = this.seconds.value-this.elapsed+'s, 100ms, 100ms'
-      }
-      this.working = true
-      document.title = this.toTimeString(this.seconds.value-this.elapsed);
-      this.timer = setInterval(() => {
-        this.elapsed = this.elapsed+1;
-        if (this.lightningMode) {
-          this.elapsed = this.elapsed+1;
-        }
-        if (this.elapsed >= this.seconds.value) {
-          if (this.autoplay) {
-            this.skip(true)
-            this.goToNextTask()
-            this.play()
-          } else {
-            this.modal = 'moreTime'
-            var audio = new Audio(this.beep);
-            audio.play();
-          }
-          clearInterval(this.timer);
-        }
-
-        document.title = this.toTimeString(this.seconds.value-this.elapsed);
-      }, 1000)
-    },
-    getLocalStorage(key) {
-      return localStorage.getItem(key)
-    },
-    setLocalStorage(key, value) {
-      if (this.uid) return;
-      return localStorage.setItem(key, value);
-    },
-    goToNextColor() {
-      const currentIndex = this.colors.indexOf(this.activeColor)
-      const nextIndex = currentIndex + 1
-      if (nextIndex >= this.colors.length) {
-        this.activeColor = this.colors[0]
-      } else {
-        this.activeColor = this.colors[nextIndex]
-      }
-    },
-    beforeWindowUnload(e) {
-      this.invokeSave()
-    },
-    createWelcome() {
-      this.title = 'Check it out'
-      this.seconds = {value: 60, label: '1 minute'}
-      this.notes = { blocks: this.welcomeBlocks, time: Date.now(), version: '2.18.0' }
-      this.elapsed = 0
-      this.completed = false
-      this.taskId = uuidv4()
-      this.activeColor = this.colors[Math.floor(Math.random() * this.colors.length)];
-      this.setLocalStorage('color', this.activeColor);
-      this.setLocalStorage('taskId', this.taskId);
-      this.setLocalStorage('title', this.title);
-      this.setLocalStorage('seconds', JSON.stringify(this.seconds));
-      this.setLocalStorage('notes', JSON.stringify(this.notes));
-      this.setLocalStorage('elapsed', this.elapsed);
-      this.setLocalStorage('completed', this.completed);
-      this.setLocalStorage('setupComplete', '1')
-      this.config.data = this.notes
-    },
-    addNewFolder(folderName) {
-      var folderId = uuidv4();
-      var newFolder = {
-        id: folderId,
-        name: folderName
-      }
-      var folders = []
-      if (this.getLocalStorage('folders')) {
-        folders = JSON.parse(this.getLocalStorage('folders'))
-      }
-      folders.push(newFolder)
-      this.setLocalStorage('folders', JSON.stringify(folders))
-      this.folders = folders
-      return newFolder
-    },
-    pause() {
-      this.working = false
-      this.transitionDuration = '0s, 100ms, 100ms'
-      this.progressBarWidth = 'calc('+this.elapsed/this.seconds.value*100+'vw - 30px)'
-      clearInterval(this.timer)
-      document.title = 'Focusmix. Free Task Timer Online for Students. Aesthetic.'
-    },
-
-    addNewTask(taskTitle, setCurrent = false) {
-      var taskId = uuidv4();
-      var newTask = {
-        id: taskId,
-        title: taskTitle,
-        completed: false,
-        elapsed: 0
-      }
-      this.tasks.push(newTask)
-      
-      if (setCurrent) {
-        this.setLocalStorage('taskId', taskId)
-        this.setLocalStorage('seconds', {value: 3600, label: '1 hour'})
-        this.setLocalStorage('completed', false)
-        this.setLocalStorage('elapsed', 0)
-        this.setLocalStorage('notes', { blocks: [{data: {text: 'Write here.'}, id: "moTtRFW4VL", type: "paragraph"}], time: Date.now(), version: '2.18.0' })
-
-        db.collection('Users').doc(this.uid).get((doc) => {
-          db.collection('Users').doc(this.uid).set({
-            workingTask: taskId
-          }, { merge: true })
-        })
-        this.changeTask(newTask)
-      }
-      db.collection('Users').doc(this.uid).collection('Tasks').doc(taskId).set({
-        title: taskTitle,
-        seconds: { value: 3600, label: '1 hour' },
-        elapsed: 0,
-        completed: false,
-        notes: { blocks: [{data: {text: 'Write here.'}, id: "moTtRFW4VL", type: "paragraph"}], time: Date.now(), version: '2.18.0' }
-      })
-      this.newTaskTitle = ''
-      return newTask
-    },
-    pullData() {
-      let pageReady = new Promise((resolve, reject) => {
-        db.collection('Users').doc(this.uid).get().then(doc => {
-          this.currentFolder = doc.data().workingFolder
-          this.activeColor = doc.data().options.color
-          this.taskId = doc.data().workingTask
-          return doc;
-        }).then((doc) => {
-          db.collection('Users').doc(this.uid).collection('Tasks').doc(this.taskId).get().then(doc => {
-            if (doc.exists) {
-              this.title = doc.data().title
-              this.seconds = doc.data().seconds
-              this.completed = Boolean(parseInt(doc.data().completed))
-              this.elapsed = doc.data().elapsed
-              this.notes = doc.data().notes
-
-              this.progressBarWidth = 'calc('+this.elapsed/this.seconds.value*100+'vw - 30px)'
-              this.transitionDuration = this.seconds.value-this.elapsed+'s, 100ms, 100ms'
-              this.config.data = this.notes
-            }
-          }).then(() => {
-            db.collection('Users').doc(this.uid).get().then((doc) => {
-              this.folders = doc.data().folders
-            }).then(() => {
-              if (this.currentFolder) {
-                db.collection('Users').doc(this.uid).collection('Folders').doc(this.currentFolder.id).get().then(doc => {
-                  var orderedTasks = doc.data().tasks;
-                  if (doc.exists) {
-                    this.tasks = orderedTasks
-                    resolve(true)
-                  }
-                }).catch(() => reject(false))
-              }
-            }).catch(() => reject(false))
-          }).catch(() => reject(false))
-        }).catch(() => reject(false))
-      })
-      pageReady.then(() => {
-        this.pageReady = true
-        window.localStorage.clear()
-      }).catch(() => {
-        this.pageReady = false
-      })
-    },
-    shuffleTasks() {
-      var original = this.tasks
-      var copy = [].concat(original);
-      for (let i = copy.length - 1; i > 0; i--) {
-        let j = Math.floor(Math.random() * (i + 1));
-        [copy[i], copy[j]] = [copy[j], copy[i]];
-      }
-      return copy
-    },
-    getNextArrItem(startIndex, arr) {
-      var nextIndex = startIndex + 1
-      if (nextIndex >= arr.length) {
-        nextIndex = false
-      }
-      return arr[nextIndex]
-    },
-    startOver() {
-      this.lightningMode = false
-      this.animation = 'none'
-      this.transitionDuration = '0.0001s, 100ms, 100ms'
-      this.working = false
-      clearInterval(this.timer);
-      this.elapsed = 0
-      this.progressBarWidth = '1px'
-    },
-    invokeSave() {
-      if (this.uid) {
-        this.setLocalStorage('seconds', JSON.stringify(this.seconds))
-        this.setLocalStorage('completed', this.completed)
-        this.setLocalStorage('elapsed', this.elapsed)
-        this.setLocalStorage('color', this.activeColor)
-
-        this.saving = true;
-        this.$refs.editor._data.state.editor.save().then((data) => {
-          this.notes = data
-          this.setLocalStorage('notes', JSON.stringify(data))
-        })
-        return;
-      }
-      db.collection('Users').doc(this.uid).set({
-        options: {
-          color: this.activeColor,
-        },
-        workingFolder: this.currentFolder,
-        workingTask: this.taskId,
-        folders: this.folders
-      }, { merge: true })
-
-      db.collection('Users').doc(this.uid).collection('Tasks').doc(this.taskId).set({
-        notes: this.notes,
-        title: this.title,
-        seconds: this.seconds,
-        elapsed: this.elapsed,
-        completed: this.completed
-      }, { merge: true })
-
-      db.collection('Users').doc(this.uid).collection('Folders').doc(this.currentFolder.id).set({
-        tasks: this.tasks,
-      }, { merge: true })
-    },
-    logout() {
-      window.localStorage.clear()
-      setTimeout(() => {
-        auth.signOut().then(() => {
-          window.location.reload()
-        });
-      }, 1111)
-    },
-    sendEmailSignIn() {
-      auth.sendSignInLinkToEmail(this.email, {
-        url: window.location.href,
-        handleCodeInApp: true
-      })
-      .then(() => {
-        localStorage.setItem('emailForSignIn', this.email);
-      });
-      this.waitingForSignin = true;
-    },
+  components: {
+    UserNav,
+    MainContent,
+    ControlsNav,
+    AboutSiteModal,
+    QueueModal,
+    TaskOverModal
   }
 }
 </script>
@@ -920,18 +56,13 @@ body {
   }
 }
 @media (min-width: 1024px) {
-  .controls {
-    width: 800px;
-  }
-}
-@media (min-width: 1024px) {
   .container, .controls {
     margin: 0 auto;
   }
 }
 .container {
   position: relative;
-  z-index: 50;
+  z-index: 10;
 }
 .codex-editor__redactor {
   height: calc(100% - 100px);
@@ -952,7 +83,7 @@ body {
   bottom: 0;
   left: 50%;
   transform: translateX(-50%);
-  padding: 0 10px 0 30px;
+  padding: 10px 10px 10px 30px;
   background: rgba(255,255,255,.95);
   z-index: 100;
   display: grid;
@@ -963,7 +94,7 @@ body {
     display: grid;
     grid-template-columns: 1fr 4fr 1fr;
     align-items: center;
-    padding: 0 10px;
+    padding: 10px 10px;
   }
 }
 @media (min-width: 1024px) {
@@ -971,7 +102,7 @@ body {
     border-radius: 100px;
     bottom: 20px;
     display: grid;
-    grid-template-columns: 1fr 2fr 1fr;
+    grid-template-columns: repeat(3, 1fr);
   }
 }
 button {
@@ -1013,7 +144,7 @@ input[type="text"]:focus,input[type="text"]:focus:not(.show), select:focus, text
 .task-details {
   font-size: 1rem;
 }
-.folders {
+.queue {
   display: flex;
   align-items: center;
   list-style: none;
@@ -1023,14 +154,18 @@ input[type="text"]:focus,input[type="text"]:focus:not(.show), select:focus, text
 .action {
   display: flex;
   align-items: center;
-  justify-content: space-around;
-  gap: 20px;
+  justify-content: flex-start;
+  gap: 30px;
 }
-.browse {
+@media (max-width: 1023px) {
+  .action {
+    justify-content: center;
+  }
+}
+.play-pause {
   display: flex;
   align-items: center;
   justify-content: space-between;
-  gap: 20px;
 }
 @media (min-width: 0px) and (max-width: 1023px) {
   .mark-complete {
@@ -1048,32 +183,43 @@ input[type="text"]:focus,input[type="text"]:focus:not(.show), select:focus, text
     align-items: center;
   }
 }
+.site-nav button {
+  position: relative;
+  z-index: 60;
+}
+.mark-complete {
+  flex: 1;
+}
 .color, .progress-bar, .progress-bar-color {
-  width: 0;
   position: fixed;
   height: 100vh;
   top: 0;
-  z-index: 0;
+  z-index: -2;
 }
 .progress-bar {
   left: 30px;
 }
 .progress-bar-color {
-  width: 1px;
+  transition-property: width, background;
+  transition-timing-function: linear;
+  transition-duration: 60s, 1ms;
+  position: fixed;
+  top: 0;
+  bottom: 0;
   left: 30px;
-  transition-property: width, background, max-width;
-  transition-timing-function: ease-in-out;
-  transition-duration: 100ms, 100ms, 100ms;
+  width: 0;
+  max-width: calc(100vw - 30px)
 }
 .color {
   left: 0;
   min-width: 30px;
-  z-index: 3;
+  z-index: -1;
 }
 .site-nav {
   z-index: 60;
   text-align: right;
   display: inline-block;
+  position: relative;
 }
 @media (min-width: 0px) and (max-width: 1023px) {
   .site-nav {
@@ -1097,7 +243,7 @@ input[type="text"]:focus,input[type="text"]:focus:not(.show), select:focus, text
 }
 .muted {
   opacity: .5;
-  transition: opacity 100ms;
+  transition: opacity 1ms;
 }
 .muted:hover {
   opacity: .75;
@@ -1140,19 +286,19 @@ input[type="text"]:focus,input[type="text"]:focus:not(.show), select:focus, text
 @media (min-width: 0px) and (max-width: 1023px) {
   .queue-list {
     left: 0;
-    bottom: 60px;
+    bottom: 0;
   }
 }
 @media (min-width: 1024px) {
   .queue-list {
-    bottom: 60px;
+    bottom: 0;
     left: 0;
   }
 }
 .change-folder {
   font-size: 17px;
 }
-.add-new-folder, .add-new-task {
+.add-new-folder {
   display: flex;
   align-items: center;
   justify-content: space-between;
@@ -1164,17 +310,12 @@ input[type="text"]:focus,input[type="text"]:focus:not(.show), select:focus, text
   bottom: 60px;
   width: 290px;
 }
-@media (min-width: 0px) and (max-width: 1023px) {
-  .add-new-task {
-    bottom: 60px;
-    width: 290px;
-  }
-}
-@media (min-width: 1024px) {
-  .add-new-task {
-    bottom: 60px;
-    width: 290px;
-  }
+.add-new-task {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: 20px;
+  padding-right: 10px;
 }
 .tooltip {
   position: relative;
@@ -1250,10 +391,11 @@ a {
   margin: 0 auto;
   z-index: 102;
   border-radius: 4px;
-  position: absolute;
+  position: fixed;
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
+  text-align: left;
 }
 button.close-button {
   font-size: 3rem;
@@ -1285,7 +427,7 @@ button[type="submit"]:not(.invisible),button[type="button"].submit {
 }
 .features li {
   padding-left: 40px;
-  background: url(./assets/images/star.svg) no-repeat center left;
+  background: url(@/assets/images/star.svg) no-repeat center left;
 }
 .saving {
   opacity: .25;
@@ -1321,6 +463,28 @@ code, .ce-code__textarea {
 @media (min-width: 0px) and (max-width: 1023px) {
   .queue-login {
     display: none;
+  }
+}
+.task-list-item {
+  margin-bottom: 5px;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+.task-list-item:not(:last-child) {
+  border-bottom: #ddd 1px solid;
+}
+.queue-modal {
+  padding-bottom: 40px;
+}
+.task-list {
+  max-height: 240px;
+  overflow: auto;
+  padding-right: 10px;
+}
+@media (max-width: 600px) {
+  .ce-popover {
+    bottom: 82px !important
   }
 }
 </style>
