@@ -1,3 +1,4 @@
-module.exports = {
-  runtimeCompiler: true
-}
+const { defineConfig } = require('@vue/cli-service')
+module.exports = defineConfig({
+  transpileDependencies: true
+})
